add initial test cases for metrics
