+++ tests/metrics/test_iou.py
@@ -1,0 +1,20 @@
+import unittest
+
+import numpy as np
+
+from zae_engine.metrics.iou import miou, giou
+
+
+class TestIoU(unittest.TestCase):
+    def setUp(self) -> None:
+        pass
+
+    def test_mIoU(self):
+        pass
+
+    def test_gIoU(self):
+        pass
+
+
+if __name__ == "__main__":
+    unittest.main()
